+++ physicsEngine/src/processable/raytest/RayTester.cpp
@@ -35,16 +35,19 @@ namespace urchin
 	void RayTester::execute(float, const Vector3<float> &)
 	{
 		std::vector<AbstractWorkBody *> bodiesAABBoxHitRay = collisionWorld->getBroadPhaseManager()->rayTest(ray);
-		std::vector<std::shared_ptr<RayCastResult<float>>> rayCastResults = collisionWorld->getNarrowPhaseManager()->rayTest(ray, bodiesAABBoxHitRay);
+		std::vector<std::shared_ptr<RayCastResult<double>>> rayCastResults = collisionWorld->getNarrowPhaseManager()->rayTest(ray, bodiesAABBoxHitRay);
 
 		std::vector<RayTestSingleResult> rayTestResults;
 		rayTestResults.reserve(rayCastResults.size());
 
-		for(const std::shared_ptr<RayCastResult<float>> &rayCastResult : rayCastResults)
+		for(const std::shared_ptr<RayCastResult<double>> &rayCastResult : rayCastResults)
 		{
 			if(rayCastResult->hasTimeOfImpactResult())
 			{
-				RayTestSingleResult rayTestSingleResult(rayCastResult->getNormal(), rayCastResult->getHitPointB(), rayCastResult->getTimeToHit());
+				RayTestSingleResult rayTestSingleResult(
+						rayCastResult->getNormal().template cast<float>(),
+						rayCastResult->getHitPointB().template cast<float>(),
+						static_cast<float>(rayCastResult->getTimeToHit()));
 				rayTestResults.push_back(rayTestSingleResult);
 			}
 		}
